requirements.txt
attrs
colorama==0.3.9
enum
